requirements.txt
absl-py==1.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==5.0.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==6.0.0
blinker==1.6.3
blis==0.7.11
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.5.7
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.16.0
colorlog==6.7.0
confection==0.1.3
contexttimer==0.3.3
crcmod==1.7
cycler==0.11.0
cymem==2.0.8
decorator==5.1.1
decord==0.6.0
diffusers==0.16.0
distlib==0.3.7
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
einops==0.4.1
fairscale==0.4.4
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
ftfy==6.1.1
functorch==1.13.1
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.22.0
google-auth-oauthlib==0.4.6
grpcio==1.56.2
h5py==3.8.0
huggingface-hub==0.16.4
identify==2.5.24
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.10
ipython==7.34.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.1
jsonschema==4.17.3
kaggle==1.5.16
kiwisolver==1.4.4
langcodes==3.3.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
more-itertools==9.1.0
munch==4.0.0
murmurhash==1.0.10
networkx==2.6.3
nltk==3.8.1
nodeenv==1.8.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
oss2==2.18.3
packaging==23.1
pandas==1.3.5
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
plotly==5.18.0
portalocker==2.7.0
pre-commit==2.21.0
preshed==3.0.9
pretrainedmodels==0.7.4
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycryptodome==3.19.0
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==8.0.1
pytorch-metric-learning==2.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
ruamel.yaml==0.15.46
ruamel.yaml.clib==0.2.8
safetensors==0.4.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.99
sentry-sdk==1.31.0
setproctitle==1.3.2
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.23.1
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminaltables==3.1.10
text-unidecode==1.3
thinc==8.2.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.10
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchaudio==0.11.0
torchinfo==1.8.0
torchvision==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.0
typer==0.9.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
virtualenv==20.24.6
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.9
weasel==0.3.3
webdataset==0.2.73
webencodings==0.5.1
Werkzeug==2.2.3
yacs==0.1.8
zipp==3.15.0
